% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sxpdatafunc.R
\name{interpolatex}
\alias{interpolatex}
\title{interpolate x based on y. Usually x is vdm, y is sigma, interpolate the required vdm@-1s, 0s and 1s}
\usage{
interpolatex(dt, target = c(0), x = "x", y = "y", grp = "grp",
  already_ordered = FALSE)
}
\arguments{
\item{dt}{a data.table}

\item{x}{x colname string}

\item{y}{y colname string}

\item{grp}{group by column string}

\item{already_ordered}{if the table is already sorted}
}
\value{
a table.table giving lowbound highbound and fitted solution
}
\description{
take a data.table and interpolate x based y level
}

