% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sxpdatafunc.R
\name{splitcol}
\alias{splitcol}
\title{take a data.table, split col into multiple columns based separator}
\usage{
splitcol(dt, col, sep1 = "_", sep2 = ":", mode = 1,
  splitcolnames = NULL)
}
\arguments{
\item{dt}{a data.table}

\item{col}{column to split, no quote}

\item{sep1}{level 1 sep1}

\item{sep2}{level 1 sep2}

\item{mode}{default 1}

\item{splitcolnames}{used in mode 0, names of splitted columns}
}
\value{
no return, inplace change to dt
}
\description{
in place change to dt
}
\examples{
mode 1: split "X:1_Y:2" into two columns X,Y, fill with 1,2;
mode 0: split "1_2" into 1 and 2, column names defined by splitcolnames if given, otherwise use V1,V2...
}

